feat: Prometheus Exporter
Remove OTLP metrics exporting in favour of Prometheus Metrics exporting
