+++ src/lib.rs
@@ -33,6 +33,7 @@ use {
 pub mod env;
 pub mod error;
 pub mod handlers;
+pub mod metrics;
 pub mod middleware;
 pub mod providers;
 pub mod relay;
@@ -116,6 +117,14 @@ pub async fn bootstap(mut shutdown: broadcast::Receiver<()>, config: Config) -> 
             .with_timeout(Duration::from_secs(5))
             .with_protocol(Protocol::Grpc);
 
+        let resource = Resource::new(vec![
+            KeyValue::new("service.name", "echo-server"),
+            KeyValue::new(
+                "service.version",
+                state.build_info.crate_info.version.clone().to_string(),
+            ),
+        ]);
+
         let tracer = opentelemetry_otlp::new_pipeline()
             .tracing()
             .with_exporter(tracing_exporter)
@@ -126,47 +135,13 @@ pub async fn bootstap(mut shutdown: broadcast::Receiver<()>, config: Config) -> 
                     .with_max_events_per_span(64)
                     .with_max_attributes_per_span(16)
                     .with_max_events_per_span(16)
-                    .with_resource(Resource::new(vec![
-                        KeyValue::new("service.name", "echo-server"),
-                        KeyValue::new(
-                            "service.version",
-                            state.build_info.crate_info.version.clone().to_string(),
-                        ),
-                    ])),
+                    .with_resource(resource.clone()),
             )
             .install_batch(opentelemetry::runtime::Tokio)?;
 
-        let metrics_exporter = opentelemetry_otlp::new_exporter()
-            .tonic()
-            .with_endpoint(grpc_url)
-            .with_timeout(Duration::from_secs(5))
-            .with_protocol(Protocol::Grpc);
+        let metrics = Metrics::new(resource);
 
-        let meter_provider = opentelemetry_otlp::new_pipeline()
-            .metrics(tokio::spawn, tokio_interval_stream)
-            .with_exporter(metrics_exporter)
-            .with_period(Duration::from_secs(3))
-            .with_timeout(Duration::from_secs(10))
-            .with_aggregator_selector(selectors::simple::Selector::Exact)
-            .build()?;
-
-        opentelemetry::global::set_meter_provider(meter_provider.provider());
-
-        let meter = opentelemetry::global::meter("echo-server");
-        let hooks_counter = meter
-            .i64_up_down_counter("registered_webhooks")
-            .with_description("The number of currently registered webhooks")
-            .init();
-
-        let notification_counter = meter
-            .u64_counter("received_notifications")
-            .with_description("The number of notification received")
-            .init();
-
-        state.set_telemetry(tracer, Metrics {
-            registered_webhooks: hooks_counter,
-            received_notifications: notification_counter,
-        })
+        state.set_telemetry(tracer, metrics)
     } else if !state.config.is_test {
         // Only log to console if telemetry disabled
         tracing_subscriber::fmt()
